Refactors Note into smaller components
Create Storybooks for all Note subcomponents
