+++ src/lib/components/notes/NoteMetadata.svelte
@@ -1,0 +1,39 @@
+<script lang="ts">
+  import { _ } from "svelte-i18n";
+  import { getUserName } from "$lib/api/accounts";
+  import type { Note, User } from "$lib/api/types";
+
+  export let note: Note;
+
+  $: user = typeof note.user === "object" ? (note.user as User) : null;
+</script>
+
+<div class="note-metadata">
+  {#if user}
+  <p class="created-by">
+    {$_("annotation.by", { values: { name: getUserName(user) } })}
+  </p>
+  {/if}
+  <p class="last-modified">
+    {#if note.updated_at && note.updated_at !== note.created_at}
+      {$_("annotation.edited", {
+        values: { date: new Date(note.updated_at).toLocaleDateString() },
+      })}
+    {:else}
+      {$_("annotation.created", {
+        values: { date: new Date(note.created_at).toLocaleDateString() },
+      })}
+    {/if}
+  </p>
+</div>
+
+<style>
+  .note-metadata {
+    display: flex;
+    align-items: baseline;
+    gap: 1rem;
+
+    font-size: var(--font-xs);
+    color: var(--gray-5);
+  }
+</style>
